Update B - Grow The Tree .cpp

+++ B - Grow The Tree .cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 #define boost		          ios_base::sync_with_stdio(0);
-using ll = long long int ;
+using ll = short short int ;
 
 int main()
 {
